+++ Examples/euroc_test.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+pathDatasetEuroc='/home/lz/hutpond/dataset/orbslam' #Example, it is necesary to change it by the dataset path
+
+#------------------------------------
+# Monocular Examples
+echo "Launching MH01 with Monocular sensor"
+./Monocular/mono_test